+++ docs/misc/changelog.rst
@@ -15,6 +15,7 @@ Breaking Changes:
 New Features:
 ^^^^^^^^^^^^^
 - Added official support for Python 3.13
+- Added Prioritized Experience Replay for DQN
 
 Bug Fixes:
 ^^^^^^^^^^